backend/requirements.txt
fastapi==0.115.12
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
alembic==1.12.1
python-dotenv==1.0.0
requests==2.32.3
PyJWT==2.9.0
httpx==0.28.1
pika==1.3.1